separate model from API app into a new file
