+++ pyproject.toml
@@ -22,5 +22,5 @@ ruff = "^0.0.287"
 ipython = "^8.15.0"
 
 [build-system]
-requires = ["poetry-core"]
+requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
